oops looks liek we have to use floats for openGL

+++ include/nbody/body.h
@@ -3,18 +3,18 @@
 
 namespace nbody{
 	class Body{
-		Vector3d _position; // use double > float; will store doubles and print floats
-		Vector3d _velocity;
-		Vector3d _acceleration;
+		Vector3f _position; 
+		Vector3f _velocity;
+		Vector3f _acceleration;
 		double _mass;
 	public:
 		Body(): _position{}, _velocity{}, _acceleration{}, _mass{} {}
-		inline Vector3d position() const { return _position; }
-		inline Vector3d& position() { return _position; }
-		inline Vector3d velocity() const { return _velocity; }
-		inline Vector3d& velocity() { return _velocity; }
-		inline Vector3d acceleration() const { return _acceleration; }
-		inline Vector3d& acceleration() { return _acceleration; }
+		inline Vector3f position() const { return _position; }
+		inline Vector3f& position() { return _position; }
+		inline Vector3f velocity() const { return _velocity; }
+		inline Vector3f& velocity() { return _velocity; }
+		inline Vector3f acceleration() const { return _acceleration; }
+		inline Vector3f& acceleration() { return _acceleration; }
 		inline float mass() const { return _mass; }
 		friend std::istream& operator>>( std::istream &is, Body &body );
 		friend std::ostream& operator<<( std::ostream &os, const Body &body );
